pyproject.toml
[project]
name = "frontier-model-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.7,<3",
  "pyyaml>=6.0.1,<7",
]

[project.scripts]
fmf = "fmf.cli:main"

[project.optional-dependencies]
aws = [
  "boto3>=1.34,<2",
]
azure = [
  "azure-identity>=1.16,<2",
  "azure-keyvault-secrets>=4.8,<5",
]
sharepoint = [
  "msgraph-sdk>=1.0.0,<2",
]
ocr = [
  "pytesseract>=0.3.10,<1",
]
delta = [
  "deltalake>=0.17,<0.21",
]
redshift = [
  "redshift-connector>=2.1,<3",
]
excel = [
  "openpyxl>=3.1,<4",
]
parquet = [
  "pyarrow>=14,<17",
]
test = [
  "pytest>=8,<9",
  "pytest-cov>=5,<6",
  "pyarrow>=14,<17",
  "moto[s3]>=5,<6",
]
